refactor(game): peel the small shared helpers out of the game commands
Seven functions that have nothing to do with starting a career but were
sitting in the middle of the file that does: error shaping for the frontend
(`first_package_error_message`, `encode_error_param`,
`map_save_manager_lock_error`), reading the active stats session, and the
three defaults a new career is created with.

They move to `game/helpers.rs` with their three tests. `mod.rs` 3972 → 3875.

Two visibility notes. `mod.rs` imports the submodule through a private glob,
so a moved item needs only `pub(super)` to stay reachable from here and from
the tests; only `first_package_error_message` and `default_save_name` are
re-exported `pub(crate)`, because only they have callers elsewhere in the
crate. And `map_save_manager_lock_error` was `pub(crate)` with no caller
outside this module at all, so it is `pub(super)` now — narrower than it was,
not wider.

`default_league_name`, `long_date_format` and `default_save_name` move
verbatim, including how oddly they are written. They build user-visible
English — "Premier Division", "%B %d, %Y", "Alex's Career" — out of joined
arrays and char-by-char pushes rather than string literals. That is what a
literal looks like when it has been shaped to get past `audit-i18n.mjs`,
which scans `src-tauri/src` for exactly those. Rewriting them plainly would
expose three strings that need translation keys in all 11 locales, so it is
its own change, not a rename. Noted here so it does not get lost.

Test list unchanged: 212 lib tests before, 212 after, leaf for leaf.
